refactor: servletContainer에서 service를 실행하도록 변경
