+++ tomcat/src/main/java/org/apache/catalina/servlet/WebServlet.java
@@ -9,7 +9,7 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface RequestMapping {
+public @interface WebServlet {
 
     String value() default "";
 }
