+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -3,14 +3,12 @@ package org.apache.coyote.http11;
 import com.techcourse.exception.UncheckedServletException;
 import java.io.IOException;
 import java.net.Socket;
-import org.apache.catalina.exception.CatalinaException;
 import org.apache.catalina.http.HttpRequest;
 import org.apache.catalina.http.HttpResponse;
 import org.apache.catalina.http.body.HttpResponseBody;
 import org.apache.catalina.http.header.HttpHeaders;
 import org.apache.catalina.http.startline.HttpResponseLine;
-import org.apache.catalina.servlet.Controller;
-import org.apache.catalina.servlet.RequestMapper;
+import org.apache.catalina.servlet.ServletContainer;
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,10 +35,9 @@ public class Http11Processor implements Runnable, Processor {
              final var outputStream = connection.getOutputStream()) {
             HttpRequest request = HttpRequest.parse(inputStream);
             HttpResponse response = createResponse(request);
-            Controller controller = RequestMapper.getController(request);
+            ServletContainer servletContainer = ServletContainer.getInstance();
+            servletContainer.service(request, response);
 
-            controller.service(request, response);
-            validateResponse(response);
             outputStream.write(response.getBytes());
             outputStream.flush();
         } catch (IOException | UncheckedServletException e) {
@@ -54,11 +51,5 @@ public class Http11Processor implements Runnable, Processor {
                 new HttpHeaders(),
                 new HttpResponseBody()
         );
-    }
-
-    private void validateResponse(HttpResponse response) {
-        if (!response.isValid()) {
-            throw new CatalinaException("Response not valid: \n" + response);
-        }
     }
 }
